requirements.txt
asgiref==3.7.2
dataclasses==0.6
Django==4.2.7
django-filter==23.3
django-js-asset==2.1.0
django-mptt==0.15.0
django-nine==0.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djongo==1.2.31
dnspython==2.4.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
Pillow==10.1.0
psycopg==3.1.8
psycopg2-binary==2.9.9
PyJWT==2.8.0
pymongo==4.6.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5